Reduce precision of shear parameters to 16 bits

+++ av1/common/mv.h
@@ -97,7 +97,7 @@ static const int trans_model_params[TRANS_TYPES] = { 0, 2, 4, 6, 6, 6, 8 };
 typedef struct {
   TransformationType wmtype;
   int32_t wmmat[8];
-  int32_t alpha, beta, gamma, delta;
+  int16_t alpha, beta, gamma, delta;
 } WarpedMotionParams;
 #endif  // CONFIG_GLOBAL_MOTION || CONFIG_WARPED_MOTION
